changed navigation to its own component
